% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejercicio1.R
\name{suma}
\alias{suma}
\title{Suma}
\usage{
suma(x, y)
}
\arguments{
\item{x}{Valor cualquiera}

\item{y}{Otro valor cualquiera}
}
\value{
Una suma
}
\description{
Sirve para sumar
}
\examples{
suma(3,2)
suma("3",2)
suma(-3,2)

}
